Remove checking for folder creation
If mkdir is successful, that means that the folders were successfully created. If they are, then are deleted (as would the subsequent check catch), then there's fundamentally a race condition with some other process that unlinked it before slurm got started.
However, this check doesn't solve the issue, as it doesn't remove the issue (the race condition could pop up before slurm startup but after the check). In fact, this event is likely if there's actually a race condition, as the two commands are fired one after another.
Thus, this slows the happy path with extra SSH requests, without actually solving the problem it's supposed to. Removing it seems the best course here