Synchronize logs using rsync, once every x seconds
The logs synchronization using the cache manager taxed heavily the server and made it eventually unresponsive as the logs became more heavy. It also was not necessary for the benchmark system, so remove that responsability from the server and instead use rsync in append-only mode. Empirically, that takes a few hundred milliseconds on my system after an hour of logs